# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Project Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Member } from './member';
import type { Project } from './project';
import type { TaskPriority } from './taskPriority';
import type { TaskStatus } from './taskStatus';

export interface Task {
  id: number;
  projectId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  assigneeId?: number | null;
  assignee?: Member;
  project?: Project;
  /** @nullable */
  dueDate?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}