# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Project Management API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputPriority } from './taskInputPriority';
import type { TaskInputStatus } from './taskInputStatus';

export interface TaskInput {
  projectId: number;
  /** @minLength 1 */
  title: string;
  description?: string;
  status?: TaskInputStatus;
  priority?: TaskInputPriority;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  dueDate?: Date | null;
}